Move some structs around, add some docs
